refactor: migration and models
